+++ sdk/eventhubs/azure_messaging_eventhubs_checkpointstore_blob/src/checkpoint_store.rs
@@ -155,7 +155,7 @@ impl BlobCheckpointStore {
 #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
 impl CheckpointStore for BlobCheckpointStore {
     /// Claims ownership of the specified partitions.
-    async fn claim_ownership(&self, ownerships: &[Ownership]) -> Result<Vec<Ownership>> {
+    async fn   claim_ownership(&self, ownerships: &[Ownership]) -> Result<Vec<Ownership>> {
         debug!("Claiming ownership for {} partitions", ownerships.len());
 
         let mut new_ownerships = Vec::new();