+++ sdk/typespec/typespec_client_core/src/error/mod.rs
@@ -3,9 +3,4 @@
 
 //! Client-specific error functions.
 
-#[cfg(feature = "http")]
-mod http_error;
-
-#[cfg(feature = "http")]
-pub use http_error::*;
 pub use typespec::error::*;