+++ sdk/typespec/typespec_client_core/CHANGELOG.md
@@ -18,8 +18,9 @@
 - Removed feature `reqwest_rustls_tls`. See [README.md](https://github.com/heaths/azure-sdk-for-rust/blob/main/sdk/typespec/typespec_client_core/README.md) for alternative HTTP client configuration.
 - Removed the `fs` module including the `FileStream` and `FileStreamBuilder` types. Moved to `examples/` to copy if needed.
 - Removed the `setters` macro.
-- Removed the cloud service specific retry headers from typespec_client_core. This change means that the `HttpError::new()` function takes an additional optional parameter which is an HTTP header which might contain an error code.
+- Removed the cloud service specific retry headers from typespec_client_core.
 - Renamed `RawResponse` to `BufResponse`. New `RawResponse` contains complete body as `Bytes` used in `ErrorKind::HttpResponse`.
+- Removed HttpError type from typespec_client_core because it is an azure_core construct.
 
 ## 0.6.0 (2025-08-01)